% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Document.R
\name{brute_force_knapsack_C}
\alias{brute_force_knapsack_C}
\title{C++ implementation of the brute force algorithm}
\usage{
brute_force_knapsack_C(x, W)
}
\arguments{
\item{x}{Same as for \link{brute_force_knapsack}, but must be a matrix.}

\item{W}{Same as for \link{brute_force_knapsack}.}
}
\value{
A list with two elements; the maximum value and the elements that are packed.
}
\description{
An improved function of \link{brute_force_knapsack}. The function is an R function wrapped around a C++ function.
}
\examples{
brute_force_knapsack_C(x = knapsack_objects[1:8,], W = 3500)
}

