% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Print the estimated coefficient values of a linreg object}
\usage{
\method{coef}{linreg}(x)
}
\arguments{
\item{x}{A linreg object.}
}
\value{
a numeric vector of  estimated coefficients values.
}
\description{
A method that prints out the estimated coefficients values of the regression model.
}
\examples{
a <- linreg(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
coef(a)
coefficients(a)
}
\seealso{
\link{coef}. The generic function from which coef.linreg is created.
}

