% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Print the estimated coefficient values of a ridgereg object}
\usage{
\method{coef}{ridgereg}(x)
}
\arguments{
\item{x}{A ridgereg object.}
}
\value{
a numeric vector of  estimated coefficients values.
}
\description{
A method that prints out the estimated coefficients values of the ridge regression model.
}
\examples{
a <- ridgereg(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda=10)
coef(a)
coefficients(a)
}
\seealso{
\link{coef}. The generic function from which coef.ridgereg is created.
}

