% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{euclidean}
\alias{euclidean}
\title{The euclidean algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number. x must be a numeric and if its not an integer the results may be wrong.}

\item{y}{A number. y must be a numeric and if its not an integer the results may be wrong.}
}
\value{
The greatest common divisor of \code{x} and \code{y}.
The euclidean algorithm calculates the greatest common divisor of its two input variables. It is explained at the Wikipedia page \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}.
}
\description{
The euclidean algorithm
}
\examples{
euclidean(123612, 13892347912)
# [1] 4
euclidean(100, 1000)
# [1] 100
}
\seealso{
\code{\link{dijkstra}} for our other function.
}

