% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Document.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Different approaches to knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{is object that contains two vectors with weights and values. Must be a data frame with two colums w and v.}

\item{W}{is maximum weight and must be a Numeric number.}
}
\value{
A list with two elements; the maximum value and the elements that are packed.
}
\description{
Three differens approaches to solve the knapsack problem: Brute force, dynamic programming and greedy approaches.
}
\examples{
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)
}

