% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Document.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{An example set of objects.}
\format{A data frame with 2,000 rows and 2 variables \code{w} and \code{v}:
\describe{
  \item{w}{containing a vector of weights}
  \item{v}{containing a vector of values}
}}
\usage{
knapsack_objects
}
\description{
A sampled set of 2,000 objects with weights and values. The weights are discrete values.
}
\keyword{datasets}

