% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{KoladaR}
\alias{KoladaR}
\alias{KoladaR-package}
\title{KoladaR package}
\source{
Kolada, see homepage.
}
\description{
Contains the function GetKolada that extracts data using the Kolada API.
}
\examples{
GetKolada(kpi_id = "N00090", municipalities = "0580")
}
\references{
Home page of the Kolada (in swedish): \url{https://www.kolada.se/}.

Documentation page for the Kolada API (in swedish): \url{http://www.kolada.se/appspecific/rka/download/api/Kolada-API-test.html}.
}
\seealso{
The GetKolada function: \link{GetKolada}.
}

