% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Document.R
\docType{package}
\name{Lab6}
\alias{Lab6}
\alias{Lab6-package}
\title{Knapsack algorithms}
\description{
Contains three algorithms for solving the knapsack problem; the brute force, dynamic programming and greedy approaches. The brute force algorithm is implemented as both R and C++ functions.
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
\references{
Information about the knapsack problem and the three approaches: \url{https://en.wikipedia.org/wiki/Knapsack_problem}.
}
\seealso{
The functions: \link{brute_force_knapsack}.
}

