% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\docType{package}
\name{Lab7}
\alias{Lab7}
\alias{Lab7-package}
\title{Linear regression package}
\description{
A package containing a linear regression function, a linear regression class and some methods for that class.
}

