% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{linreg}
\alias{linreg}
\title{Linear regression}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula on the form response~first_predictor + second_predictor + ... . The response variable and all predictors must be in the data set provided.}

\item{data}{The data set from which to choose the predictors and the response variable in the formula.}
}
\value{
An object of S3-type class "linreg": A list of 15 elements, including fitted values and residuals.
}
\description{
This is our own created linear regression function
}
\details{
The QR-decomposition was used  to estimated the regression coefficient. Model for using linreg formula is respons(numeric)~predicted variables
}
\examples{
linreg(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
}
\seealso{
The methods that use the linreg class: \link{print.linreg}, \link{plot.linreg}, \link{summary.linreg}, \link{coef.linreg}, \link{residuals.linreg} and \link{predict.linreg}.
}

