% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Print the fitted values of a linreg object}
\usage{
\method{predict}{linreg}(x)
}
\arguments{
\item{x}{A linreg object.}
}
\value{
a numeric vector of  fitted values.
}
\description{
A method that prints out the fitted values of the regression model.
}
\examples{
a <- linreg(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
predict(a)
}
\seealso{
\link{predict}. The generic function from which predict.linreg is created.
}

