% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Print the fitted values of a ridgereg object}
\usage{
\method{predict}{ridgereg}(x, x_values = "default")
}
\arguments{
\item{x}{A ridgereg object.}

\item{x_values}{Optional. If included, the predictions are made on these x_values instead.}
}
\value{
a numeric vector of  fitted values.
}
\description{
A method that prints out the fitted values of the ridge regression model, possibly for new data.
}
\examples{
a <- ridgereg(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda=10)
predict(a) # The fitted values
df <- data.frame(x1=runif(n = 150, min = -1.5, max = 1.5), x2=runif(n = 150, min = -2, max = 2.5) )
predict(x=a, x_values = df) # Predicted on the new x-values in the df data.frame.
}
\seealso{
\link{predict}. The generic function from which predict.ridgereg is created.
}

