% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{print.linreg}
\alias{print.linreg}
\title{Print a linreg object}
\usage{
\method{print}{linreg}(x)
}
\arguments{
\item{x}{A linreg object.}
}
\value{
a list of 2 elements, Formula and estimated coefficients.
}
\description{
A method that print out the regression model and it's estimated coefficients nicely.
}
\examples{
a <- linreg(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
print(a)
}
\seealso{
\link{print}. The generic function from which print.linreg is created.
}

