% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print a ridgereg object}
\usage{
\method{print}{ridgereg}(x)
}
\arguments{
\item{x}{A ridgereg object.}
}
\value{
a list of 2 elements, Formula and estimated coefficients.
}
\description{
A method that print out the ridge regression model and it's estimated coefficients nicely.
}
\examples{
a <- ridgereg(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda=10)
print(a)
}
\seealso{
\link{print}. The generic function from which print.ridgereg is created.
}

