% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Print residuals of a linreg object}
\usage{
\method{residuals}{linreg}(x)
}
\arguments{
\item{x}{A linreg object.}
}
\value{
a numeric vector of  residuals.
}
\description{
A method that print out the residulas for the regression model.
}
\examples{
a <- linreg(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
residuals(a)
resid(a)
}
\seealso{
\link{residuals}. The generic function from which residuals.linreg is created.
}

