% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summarize the regression model of a linreg object}
\usage{
\method{summary}{linreg}(x)
}
\arguments{
\item{x}{A linreg object.}
}
\value{
a list of 3 elements.
}
\description{
A method that prints a summary of the regression model.
}
\details{
The method prints a table of the coefficients in the model; their estimate, standard error and the t and p values of the estimate.
The residual standard error and the residual degrees of freedom in the model are also presented.
}
\examples{
a <- linreg(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
summary(a)
}
\seealso{
\link{summary}. The generic function from which summary.linreg is created.
}

