% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Airport delays}
\usage{
visualize_airport_delays()
}
\value{
A plot displaying the average delay for flights in 105 airports.
}
\description{
A visualisation of flight delays in 105 airports in the U.S.A.
}
\details{
Uses information from 336,776 flights from three airports.
}
\examples{
visualize_airport_delays()
}

