% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{An example graph from Wikipedia.}
\format{A data frame with 18 rows and 3 variables \code{v1}, \code{v2} and \code{w}:
\describe{
  \item{v1}{containing a vector of start nodes}
  \item{v2}{containing a vector of end nodes}
  \item{w}{containing the weight of the edges between \code{v1} and \code{v2}}
}}
\source{
\url{https://upload.wikimedia.org/wikipedia/commons/5/57/Dijkstra_Animation.gif}
}
\usage{
wiki_graph
}
\description{
A dataset containing egde from a start node to it's neighbour nodes, from a wikipedia page (\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}) as an example of how Dijkstra's algorithm works.
}
\seealso{
\code{\link{dijkstra}} for the function to which this example graph is created.
}
\keyword{datasets}

