% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\docType{package}
\name{Lab7}
\alias{Lab7}
\alias{Lab7-package}
\title{Linear regression and Rigde regression package}
\description{
A package containing both linear and ridge regression functions, linear and ridge regression classes and some methods for these classes.
}

