% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dokumentation.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge regression}
\usage{
ridgereg(formula, data, lambda = 0)
}
\arguments{
\item{formula}{A formula on the form response~first_predictor + second_predictor + ... . The response variable and all predictors must be in the data set provided.}

\item{data}{The data set from which to choose the predictors and the response variable in the formula.}

\item{lambda}{The value of lambda, lambda = 0 as default which is equals to OLS.}
}
\value{
An object of S3-type class "ridgereg": A list of nine elements, including fitted values and residuals.
}
\description{
This is our own created ridge regression function
}
\details{
This is a linear regression with one extra shrinkage parameter(lambda).
Note that the data should be normalized before using this function.
}
\examples{
ridgereg(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda=10)
}
\seealso{
The methods that use the ridgereg class: \link{print.ridgereg}, \link{coef.ridgereg} and \link{predict.ridgereg}.
}

